requirements.txt
rich>=13.0.0
Pillow>=9.0.0
requests>=2.25.0
flask>=2.0.0
pyyaml>=6.0
numpy>=1.23.0
opencv-python>=4.8.0
PyQt5>=5.15.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
